All Factors exercise - completed
I'd like to have a look at these with something more efficient (currently iterating through a loop for every number up until given value), but that's for Future MK to look at. I know, for example, that a factor can only be a maximum of 50% of the value - then 33%, 25%, 20%, etc. - but I'll spend some time figuring that one out. I think maybe checking if number/x is a whole number until x > (number/2) or something?